# src/docs/management/management_swagger.yaml
basePath: /api/v1
definitions:
...
  api.CacheCreateRequest:
    properties:
      name:
        type: string
      port:
        type: integer
      retention:
        type: integer
      token:
        type: string
    required:
    - name
    - port
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  model.Agent:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      token:
        type: string
      updatedAt:
        type: string
      uuid:
        type: string
      workspace_id:
        type: integer
    type: object
  model.BinaryCache:
    properties:
      access:
        $ref: '#/definitions/model.BinaryCacheAccess'
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      port:
        type: integer
      publicKey:
        type: string
      retention:
        type: integer
      store_paths:
        items:
          $ref: '#/definitions/model.StorePath'
        type: array
      token:
        type: string
      updatedAt:
        type: string
      url:
        type: string
      uuid:
        type: string
      workspaces:
        items:
          $ref: '#/definitions/model.Workspace'
        type: array
    type: object
  model.BinaryCacheAccess:
    enum:
    - private
    - public
    type: string
    x-enum-varnames:
    - Private
    - Public
...
  model.ErrorResponse:
    properties:
      additional_info:
        items:
          type: string
        type: array
      error:
        example: not found
        type: string
    type: object
  model.StorePath:
    properties:
      binaryCacheId:
        description: |-
          BinaryCacheId uint         `json:"binarycacheId"`
          BinaryCache   *BinaryCache `json:"-"`
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      deriver:
        type: string
      fileHash:
        type: string
      fileSize:
        type: integer
      id:
        type: integer
      narHash:
        type: string
      narSize:
        type: integer
      references:
        description: Stored as a string (serialized)
        type: string
      storeHash:
        type: string
      storeSuffix:
        type: string
      updatedAt:
        type: string
    type: object
  model.Workspace:
    properties:
      agents:
        items:
          $ref: '#/definitions/model.Agent'
        type: array
      binary_cache:
        $ref: '#/definitions/model.BinaryCache'
      binary_cache_id:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      token:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
  description: API for managing binary caches, workspaces, and agents.
  title: Management API
  version: "1.0"
paths:
  /cache:
    post:
      consumes:
      - application/json
      description: Create a new binary cache.
      parameters:
      - description: Cache details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CacheCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.BinaryCache'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Create binary cache
      tags:
      - cache
  /cache/{name}:
    get:
      description: Get detailed information about a binary cache.
      parameters:
      - description: Cache Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get cache info
      tags:
      - cache
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"